# src/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: E-Waste Workflow API
  version: 2.0.0
  description: Sprint 1 authentication/health and Sprint 2 donor-to-handoff workflow contract.
servers:
  - url: http://localhost:8080
    description: Local development and test server
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate a user and create an active session
      operationId: login
      security: []
      description: Invalid, unknown, and disabled credentials return the same generic error.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: donor1@ewaste.test
              password: example-password
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/InvalidCredentials'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate the access and refresh JWT pair
      operationId: refresh
      security: []
      description: The refresh JWT is sent in JSON and is rotated after successful use.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token pair rotated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/InvalidSession'}
        '429': {$ref: '#/components/responses/RateLimited'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the current authenticated session
      operationId: logout
      description: Requires an active access JWT. No request body is required.
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '204':
          description: Session revoked successfully
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/batches:
    get:
      tags: [Batches]
      summary: List batches visible to the authenticated actor
      operationId: listBatches
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [DRAFT, SUBMITTED, MATCHED, APPROVED, ASSIGNED, COLLECTED, FAILED_COLLECTION]
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Scoped paginated batch list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchListResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    post:
      tags: [Batches]
      summary: Create a donor batch draft
      operationId: createBatchDraft
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDraftRequest'
            example:
              category: laptops
              quantity: 10
              estimated_weight_kg: 25.50
              condition_rating: reusable
              is_data_bearing: true
              zone: central
              collection_deadline: '2026-09-23T02:00:00.000000Z'
              notes: Stored in the loading bay.
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/batches/{batch_id}:
    parameters:
      - $ref: '#/components/parameters/BatchId'
    get:
      tags: [Batches]
      summary: Get one scoped batch
      operationId: getBatch
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Batches]
      summary: Edit a donor-owned draft batch
      operationId: editBatchDraft
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDraftRequest'
      responses:
        '200':
          description: Draft updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/batches/{batch_id}/submit:
    post:
      tags: [Batches]
      summary: Validate and submit a donor draft
      operationId: submitBatch
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      responses:
        '200':
          description: Batch submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/batches/{batch_id}/receipt:
    post:
      tags: [Batches]
      summary: Verify the facility-received batch
      operationId: verifyReceipt
      description: |
        A permitted RECYCLER records the authoritative facility receipt. The command is valid
        only for a COLLECTED batch with an accepted claim and assignment in the caller's
        processing-facility organisation. It advances COLLECTED to VERIFIED atomically with the
        receipt row, audit event, and ReceiptVerified outbox record. Final analytics anomaly_codes
        are not returned by this command.
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptRequest'
      responses:
        '200':
          description: Receipt verified and batch advanced to VERIFIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/opportunities:
    get:
      tags: [Opportunities]
      summary: List eligible matching opportunities
      operationId: listOpportunities
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Scoped eligible opportunities, or an empty list while matching is in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityListResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/opportunities/{batch_id}:
    parameters:
      - $ref: '#/components/parameters/BatchId'
    get:
      tags: [Opportunities]
      summary: Get one eligible matching opportunity
      operationId: getOpportunity
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Opportunity detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/batches/{batch_id}/claim:
    post:
      tags: [Opportunities]
      summary: Claim one eligible opportunity using FCFS transaction controls
      operationId: claimOpportunity
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
            example:
              expected_version: 3
              claim_epoch: '1'
              notes: Ready for collection.
      responses:
        '200':
          description: Claim confirmed and batch approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/batches/{batch_id}/assignments:
    post:
      tags: [Assignments]
      summary: Select an approved batch and create an accepted assignment
      operationId: selectAssignment
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentSelectionRequest'
            example:
              expected_version: 4
              claim_epoch: '1'
              collector_scope_id: 9f6d5c3a-37e1-4e0e-a5f6-0f7f4e2b2c99
      responses:
        '201':
          description: Assignment created and batch assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/assignments:
    get:
      tags: [Assignments]
      summary: List assignments visible to the authenticated collector
      operationId: listAssignments
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACCEPTED, COMPLETED, FAILED, SUPERSEDED]
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Scoped assignment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentListResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/assignments/{assignment_id}:
    parameters:
      - $ref: '#/components/parameters/AssignmentId'
    get:
      tags: [Assignments]
      summary: Get one scoped assignment
      operationId: getAssignment
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Assignment detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/assignments/{assignment_id}/accept:
    post:
      tags: [Assignments]
      summary: Accept an assignment (legacy compatibility command)
      operationId: acceptAssignment
      deprecated: true
      description: New selection creates an ACCEPTED assignment. This endpoint remains for the legacy UI contract only.
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      responses:
        '200':
          description: Assignment accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/v1/assignments/{assignment_id}/reject:
    post:
      tags: [Assignments]
      summary: Reject an assignment before pickup
      operationId: rejectAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectAssignmentRequest'
            example:
              rejection_reason: route_unavailable
      responses:
        '200':
          description: Assignment rejected and batch returned to APPROVED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/assignments/{assignment_id}/handoff:
    post:
      tags: [Assignments]
      summary: Record a successful collection handoff
      operationId: recordCollectionHandoff
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoffRequest'
            example:
              pickup_occurred_at: '2026-09-19T02:00:00.000000Z'
              donor_representative_name: Representative
              actual_item_count: 10
              verification_hash: 0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef
              notes: Items received.
      responses:
        '200':
          description: Handoff recorded and batch collected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/assignments/{assignment_id}/fail:
    post:
      tags: [Assignments]
      summary: Report a failed pickup
      operationId: reportFailedPickup
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailedPickupRequest'
            example:
              failure_reason: collector_absent
              observed_details: No authorised recipient was present.
      responses:
        '200':
          description: Failed pickup recorded and assignment closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentMutationResponse'
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/batches/{batch_id}/treatment:
    post:
      tags: [Processing]
      summary: Record a facility treatment outcome
      operationId: recordTreatment
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TreatmentRequest'}
      responses:
        '200': {description: Treatment recorded and batch advanced to RECYCLED, content: {application/json: {schema: {$ref: '#/components/schemas/TreatmentMutationResponse'}}}}
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /api/v1/batches/{batch_id}/evidence:
    post:
      tags: [Evidence]
      summary: Upload optional receipt or treatment evidence after synchronous D2 validation
      operationId: uploadEvidence
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, lifecycle_stage]
              properties:
                file: {type: string, format: binary}
                lifecycle_stage: {type: string, enum: [RECEIPT, TREATMENT]}
      responses:
        '201': {description: Evidence metadata saved after synchronous D2 ingestion validation. Invalid type or size returns 422 and does not persist an evidence row., content: {application/json: {schema: {$ref: '#/components/schemas/EvidenceResponse'}}}}
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/batches/{batch_id}/evidence/{evidence_id}:
    get:
      tags: [Evidence]
      summary: Download authorised evidence
      operationId: downloadEvidence
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/CorrelationId'
        - name: evidence_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: X-Admin-Access-Reason
          in: header
          required: false
          description: Required for System Administrator downloads and recorded in the audit event; ignored for other authorised roles.
          schema: {type: string, minLength: 1, maxLength: 256}
      responses:
        '200':
          description: Validated evidence bytes. Downloads are authorised by the owning facility, Auditor, or System Administrator policy.
          content:
            application/pdf: {schema: {type: string, format: binary}}
            image/jpeg: {schema: {type: string, format: binary}}
            image/png: {schema: {type: string, format: binary}}
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
  /api/v1/batches/{batch_id}/analytics-results:
    post:
      tags: [Analytics]
      summary: Acknowledge a durable deterministic analytics result
      operationId: acknowledgeAnalyticsResult
      description: Only the authenticated analytics service identity may call this operation. Authorization must carry the configured analytics service bearer token; user access tokens are not accepted.
      parameters:
        - $ref: '#/components/parameters/BatchId'
        - $ref: '#/components/parameters/AnalyticsAuthorization'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AnalyticsAcknowledgement'}
      responses:
        '200': {description: Analytics result saved or identical replay returned, content: {application/json: {schema: {$ref: '#/components/schemas/CompletionMutationResponse'}}}}
        '401': {$ref: '#/components/responses/InvalidSession'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  parameters:
...
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Client correlation ID; generated by the server when omitted.
      schema: {type: string, minLength: 1, maxLength: 128}
...
  headers:
    CorrelationId:
      description: Correlation ID for tracing the request.
      schema: {type: string, minLength: 1, maxLength: 128}
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, format: password, minLength: 1}
...
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, refresh_expires_in]
      properties:
        access_token:
          type: string
          description: Short-lived access JWT for protected APIs.
        refresh_token:
          type: string
          description: Longer-lived refresh JWT accepted only by /auth/refresh.
        token_type: {type: string, enum: [Bearer]}
        expires_in:
          type: integer
          format: int64
          minimum: 1
          description: Access-token lifetime in seconds.
        refresh_expires_in:
          type: integer
          format: int64
          minimum: 1
          description: Refresh-token lifetime in seconds.
...
  responses:
    BadRequest:
      description: Missing or malformed request.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example:
            code: INVALID_REQUEST
            message: invalid request
            correlation_id: corr-001
      headers:
        X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}
    InvalidCredentials:
      description: Generic authentication failure; user existence is not disclosed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example:
            code: AUTH_INVALID_CREDENTIALS
            message: invalid credentials
            correlation_id: corr-login-001
      headers:
        X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}
...
    RateLimited:
      description: Too many login or refresh requests from the client address.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
          example:
            code: AUTH_RATE_LIMITED
            message: too many requests
            correlation_id: corr-rate-001
      headers:
        X-Correlation-ID: {$ref: '#/components/headers/CorrelationId'}
...